+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds an open data licence / terms of use.
+  [msom]
+
 1.15.1 (2017-07-03)
 ~~~~~~~~~~~~~~~~~~~